pyproject.toml
[project]
name = "manus-agent"
version = "0.1.0"
description = "Autonomous AI agent inspired by Manus.im with LangGraph orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "duckduckgo-search>=6.0.0",
    "playwright>=1.48.0",
    "paramiko>=3.5.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    # Phase 1 - Browser Automation
    "browser-use>=0.1.40",
    # Phase 1 - Multi-Search
    "googlesearch-python>=1.3.0",
    "baidusearch>=1.0.3",
    # Phase 1 - Web Crawling
    "crawl4ai>=0.6.3",
    "beautifulsoup4>=4.12.0",
    "html2text>=2024.2.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
